Implement an authority and use its generated pseudonyms to link to episodes.
Now validate a patient account by creating an endorsement thus linking a patient and an authority. Each project is also linked to an authority and all episodes are created with a project-specific pseudonym as before but also an authority-derived pseudonym. Thus, validated (endorsed) user accounts can find the episodes created on an opt-out basis.