Tidy up API, adding getEpisodes() for patient.
This finds all of the episodes for a patient, including explicit opt-in (from registrations) and those created by projects linked to an authority that has endorsed (validated) the patient record. Endorsement acts as a link to an authority but the link from patient to episodes is using an encrypted authority-scoped pseudonym.