Add more complex unit testing to demonstrate asymmetric encryption/decryption.
Here, given an email, identifier and date of birth, a project can send potentially send a secure message to the patient by using their stored public key. The patient, on logging in, can decrypt and read that message. However, no-one with only access to the database without the patient identifiable information can find out what a patient is linked to or read those messages.