Add skeleton model of consent and permission system.
Here a consent form is built of multiple ordered items. Consent forms are versioned and are in draft, final or deprecated. A consent item is defined by a BEHAVIOUR which is project-specific. Patients will complete a set of permissions (PermissionForm made up of PermissionItems) based on those consent forms. Some basic unit tests added. Much validation and more tests required.