Enforce rules regarding changing status of a consent form.
A DRAFT form can have anything done to it. A FINAL form can only be deprecated. A DEPRECATED form can’t have anything done to it.

This is essential in enforcing other rules relating to consent forms particularly in relation to having confidence that the contents of a consent form cannot be edited after being shown to a patient.